+++ ocdsmerge/__init__.py
@@ -1,1 +1,1 @@
-from .merge import merge, merge_versioned, get_merge_rules, WARN, RAISE, IGNORE  # noqa
+from .merge import merge, merge_versioned, get_merge_rules, RAISE, WARN, APPEND, MERGE_BY_POSITION  # noqa