+++ tests/test_collisions.py
@@ -1,24 +1,19 @@
-import json
 import os
-import warnings
 
 import pytest
 
-from ocdsmerge import merge, WARN, RAISE, IGNORE
+from ocdsmerge import merge, WARN, RAISE, APPEND, MERGE_BY_POSITION
 from ocdsmerge.merge import DuplicateIdValueError, DuplicateIdValueWarning
-from tests import read
+from tests import load
 
 
-def custom_warning_formatter(message, category, filename, lineno, line=None):
-    return '{}: {}'.format(category.__name__, str(message).replace(os.getcwd() + os.sep, ''))
+releases = load(os.path.join('schema', 'identifier-merge-duplicate-id.json'))
 
-
-warnings.formatwarning = custom_warning_formatter
-releases = json.loads(read(os.path.join('schema', 'identifier-merge-duplicate-id.json')))
 
 def test_default():
-    with pytest.warns(DuplicateIdValueWarning):
+    with pytest.warns(None) as record:
         merge(releases)
+        assert not record, 'unexpected warning: {}'.format(record[0].message)
 
 
 def test_warn():
@@ -26,14 +21,46 @@ def test_warn():
         merge(releases, collision_behavior=WARN)
 
 
-def test_ignore():
-    with pytest.warns(None) as record:
-        merge(releases, collision_behavior=IGNORE)
-        assert not record, 'unexpected warning: {}'.format(record[0].message)
-
-
 def test_raise():
     with pytest.raises(DuplicateIdValueError) as excinfo:
         merge(releases, collision_behavior=RAISE)
 
     assert str(excinfo.value) == "Multiple objects have the `id` value '1' in the `identifierMerge` array"
+
+
+def test_raise_precedence():
+    with pytest.raises(DuplicateIdValueError) as excinfo:
+        merge(releases, collision_behavior=RAISE | WARN)
+
+    assert str(excinfo.value) == "Multiple objects have the `id` value '1' in the `identifierMerge` array"
+
+
+def test_merge_by_position():
+    with pytest.warns(None) as record:
+        compiled_release = merge(releases + releases, collision_behavior=MERGE_BY_POSITION)
+        assert not record, 'unexpected warning: {}'.format(record[0].message)
+
+    assert compiled_release == load(os.path.join('schema', 'identifier-merge-duplicate-id-by-position.json'))
+
+
+def test_append():
+    with pytest.warns(None) as record:
+        compiled_release = merge(releases + releases, collision_behavior=APPEND)
+        assert not record, 'unexpected warning: {}'.format(record[0].message)
+
+    assert compiled_release == load(os.path.join('schema', 'identifier-merge-duplicate-id-append.json'))
+
+
+def test_merge_precedence():
+    with pytest.warns(None) as record:
+        compiled_release = merge(releases + releases, collision_behavior=APPEND | MERGE_BY_POSITION)
+        assert not record, 'unexpected warning: {}'.format(record[0].message)
+
+    assert compiled_release == load(os.path.join('schema', 'identifier-merge-duplicate-id-append.json'))
+
+
+def test_mixed():
+    with pytest.warns(DuplicateIdValueWarning):
+        compiled_release = merge(releases + releases, collision_behavior=MERGE_BY_POSITION | WARN)
+
+    assert compiled_release == load(os.path.join('schema', 'identifier-merge-duplicate-id-by-position.json'))